+++ packages/iov-bcp-types/src/connection.ts
@@ -196,7 +196,7 @@ export interface BcpConnection {
   readonly chainId: () => ChainId;
   readonly height: () => Promise<number>;
   readonly getTicker: (ticker: TokenTicker) => Promise<BcpQueryEnvelope<BcpTicker>>;
-  readonly getAllTickers: () => Promise<BcpQueryEnvelope<BcpTicker>>;
+  readonly getAllTickers: () => Promise<ReadonlyArray<BcpTicker>>;
 
   // accounts
   /**