+++ packages/iov-bns/types/bnsconnection.d.ts
@@ -32,7 +32,7 @@ export declare class BnsConnection implements BcpAtomicSwapConnection {
     status(): Promise<StatusResponse>;
     postTx(tx: PostableBytes): Promise<PostTxResponse>;
     getTicker(ticker: TokenTicker): Promise<BcpQueryEnvelope<BcpTicker>>;
-    getAllTickers(): Promise<BcpQueryEnvelope<BcpTicker>>;
+    getAllTickers(): Promise<ReadonlyArray<BcpTicker>>;
     getAccount(query: BcpAccountQuery): Promise<BcpAccount | undefined>;
     getNonce(query: BcpAddressQuery | BcpPubkeyQuery): Promise<Nonce>;
     /**