+++ packages/iov-ethereum/types/ethereumconnection.d.ts
@@ -17,7 +17,7 @@ export declare class EthereumConnection implements BcpConnection {
     height(): Promise<number>;
     postTx(bytes: PostableBytes): Promise<PostTxResponse>;
     getTicker(_: TokenTicker): Promise<BcpQueryEnvelope<BcpTicker>>;
-    getAllTickers(): Promise<BcpQueryEnvelope<BcpTicker>>;
+    getAllTickers(): Promise<ReadonlyArray<BcpTicker>>;
     getAccount(query: BcpAccountQuery): Promise<BcpAccount | undefined>;
     getNonce(query: BcpAddressQuery | BcpPubkeyQuery): Promise<Nonce>;
     getBlockHeader(height: number): Promise<BlockHeader>;