+++ packages/iov-lisk/src/liskconnection.spec.ts
@@ -126,10 +126,10 @@ describe("LiskConnection", () => {
 
   it("can get all tickers", async () => {
     const connection = new LiskConnection(dummynetBase, dummynetChainId);
-    const response = await connection.getAllTickers();
-    expect(response.data.length).toEqual(1);
-    expect(response.data[0].tokenTicker).toEqual("LSK");
-    expect(response.data[0].tokenName).toEqual("Lisk");
+    const tickers = await connection.getAllTickers();
+    expect(tickers.length).toEqual(1);
+    expect(tickers[0].tokenTicker).toEqual("LSK");
+    expect(tickers[0].tokenName).toEqual("Lisk");
   });
 
   it("can get chain ID", async () => {